requirements.txt
aiohttp==3.9.2
async-timeout<5.0
awscli
asyncio==3.4.3
Flask<2.4
Flask-Login>=0.5.0
gunicorn>=20.1.0
Flask-WTF>=1.0.1
Flask-SQLAlchemy>=2.5.1
GeoAlchemy2>=0.11.1 
geojson
marshmallow-sqlalchemy
watchdog>=0.9.0
lxml==4.8.0
bsddb3==6.2.7
Flask-Mail==0.9.0
Flask-Script==2.0.5
feedparser==6.0.1
sh==1.09
passlib>=1.7.4
requests>=2.20.1
simplekv
Werkzeug<3.0.0
thredds_crawler>=1.5.4
#-e git+https://github.com/pallets-eco/flask-session@a9d001ae9b40268c9189c468bb136e535d3358e1#egg=flask-session
Flask-Session
pyyaml>=6.0
#rq==1.3.0
rq
redis
rq-dashboard==0.3.4
-e git+https://github.com/ioos/ioos_qc@0b1406b52872e564885ccff5808315cf88ae28ee#egg=ioos_qc
netCDF4>=1.5.8
cf-units>=2.1.5
# cftime needs to be frozen for utime.  Some scripts could be rewritten to use
# newer functions
cftime==1.4.1
flask-cors==3.0.9
compliance-checker>=5.0.0
cc-plugin-glider>=2.0.3
#-e git+https://github.com/flasgger/flasgger.git@master#egg=flasgger
flasgger==0.9.7b2
multidict
influxdb>=5.3.1
quantities==0.14.1
